# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ACPBook Doc",
        "description": "Documentação api AcpBook",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tag": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Listar todas as tags",
                "operationId": "a52ca92f98c603351f1f68f4a2243f88",
                "parameters": [
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Filtrar tags pela descrição",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tags"
                    }
                }
            }
        },
        "/api/tag/{id}/show": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Exibir uma tag específica",
                "operationId": "6f104441e798f8c6e6378e243465b743",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da tag"
                    },
                    "404": {
                        "description": "Tag não encontrada"
                    }
                }
            }
        },
        "/api/tag/store": {
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Criar nova tag",
                "operationId": "e5b5f0f831fbad7da5a70968cfe2cb4c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "bg_color"
                                ],
                                "properties": {
                                    "description": {
                                        "description": "Descrição da tag",
                                        "type": "string"
                                    },
                                    "bg_color": {
                                        "description": "Cor de fundo da tag",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag criada com sucesso"
                    }
                }
            }
        },
        "/api/tag/{id}/update": {
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Atualizar uma tag",
                "operationId": "bab44e5742743a5f9998f9d729704a00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tag a ser atualizada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "bg_color"
                                ],
                                "properties": {
                                    "description": {
                                        "description": "Descrição da tag",
                                        "type": "string"
                                    },
                                    "bg_color": {
                                        "description": "Cor de fundo da tag",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag atualizada com sucesso"
                    },
                    "404": {
                        "description": "Tag não encontrada"
                    }
                }
            }
        },
        "/api/tag/{id}/delete": {
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Deletar uma tag",
                "operationId": "07af5ac764279b74c649ce2a198802dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tag a ser deletada",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag deletada com sucesso"
                    },
                    "404": {
                        "description": "Tag não encontrada"
                    }
                }
            }
        },
...
        "/api/user/{id}/show": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Exibir detalhes de um usuário",
                "operationId": "c8ce8fc859ded3e53e4fcd0f7438cd00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do usuário"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        },
        "/api/user/{id}/update": {
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar dados do usuário",
                "operationId": "3a271c1ec85bbf8949db575d1ea60503",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "403": {
                        "description": "Ação não autorizada"
                    }
                }
            }
        },
        "/api/user/{id}/delete": {
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Deletar um usuário",
                "operationId": "0bc5f122a227b357281aa34268d099fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "403": {
                        "description": "Ação não autorizada"
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}